+++ app/controllers/api/v0/forecasts_controller.rb
@@ -1,13 +1,5 @@
 class Api::V0::ForecastsController < ApplicationController
   def show
-    coordinates = MapquestFacade.get_lat_lon(params[:location])
-
-    
-
-    # data = MapquestFacade.get_lat_lon(params[:location])
-    # lat = data[:results].first[:locations].first[:latLng][:lat]
-    # lon = data[:results].first[:locations].first[:latLng][:lng]
-    # forecast = WeatherFacade.get_forecast(lat,lon)
-    # render json: WeatherSerializer.format_forecast(forecast)
+    render json: WeatherFacade.get_forecast(params[:q])
   end
 end
