+++ app/facades/mapquest_facade.rb
@@ -1,8 +1,7 @@
 class MapquestFacade
   def self.get_lat_lon(location)
     data = MapquestService.get_lat_lon(location)
-    require 'pry'; binding.pry
-    coordinates = "#{data[:results].first[:locations].first[:latLng][:lat]}, ##{data[:results].first[:locations].first[:latLng][:lng]}"
+    coordinates = "#{data[:results].first[:locations].first[:latLng][:lat]},#{data[:results].first[:locations].first[:latLng][:lng]}"
     split_coordinates = coordinates.split(",")
   end
 end
