+++ app/services/weather_service.rb
@@ -1,6 +1,8 @@
 class WeatherService
   def self.conn
-    Faraday.new(url: 'http://api.weatherapi.com/v1/forecast.json')
+    Faraday.new(url: 'http://api.weatherapi.com/v1/forecast.json') do |faraday|
+      faraday.headers['key'] = Rails.application.credentials.weather_api[:key]
+    end
   end
   
   def self.get_url(url)
@@ -9,6 +11,6 @@ class WeatherService
   end
 
   def self.get_forecast(lat, lon)
-    get_url("?key=#{Rails.application.credentials.weather_api[:key]}&q=#{lat},#{lon}&aqi=no&days=5")
+    get_url("?q=#{lat},#{lon}&aqi=no&days=5")
   end
 end
